requirements.txt
torch>=1.9.0
torchvision>=0.10.0
timm>=0.6.0
Flask>=2.0.0
flask-cors>=3.0.0
Pillow>=8.0.0
pyngrok>=5.0.0
numpy>=1.21.0
streamlit>=1.28.0
plotly>=5.0.0